Retter opp etter code-review

+++ web/src/main/frontend/js/felles/utils/date-utils.js
@@ -1,6 +1,6 @@
 export const formats = {
     date: {
-        short: {
+        numeric: {
             day: "2-digit",
             month: "2-digit",
             year: "numeric"
